Update ~/.xresources and ~/.xsession to work with Ubuntu
This configuration is derived cloning what you get from an
auto-configuration done by Ubuntu desktop, since it's pretty nice I
tried to replicate the same configuration in i3.

To do that I needed (seems like it) a compositor (compton) and a
couple of tricks to properly configure the pixel scale for the retina
display.

The result is not as good as Ubuntu desktop but it's better than
before.